Don't authenticate GitHubConnection in __init__() to keep it testable
To simplify the usage of dynamic connections, I called the authentication
methods of the GitHubConnection inside the __init__() (so no additional
calls are necessary). Unfortunately, this reduces the testability of
this class. So, I defined a new init() method which should then be
implemented by all connection classes and is called during the creation
of dynamic connections.