Always delete outdated repositories
Previously, we only deleted repositores explicitly for a GitHub delete
event (repository, installation). However, if we miss such an event, we
will never delete this repository and always try to scrape it, although
it is not accessible anylonger. Even a fresh scraper run won't help,
because the repository is still in Elasticsearch and, in a normal case,
we won't receive a second deletion event for this repository from GitHub.