Build Gerrit web URLs based on the "gitweb type" (cgit/gitweb)
Unfortunately, the Gerrit web URLs (used for browsing directories and
showing files) are completely different and depending on the tool
that is used for the Gerrit "web endpoint".
This change supports the usage of cgit and gitweb as Gerrit web endpoint
and builds the URLs accordingly.

I've discovered a third tool called "gitiles", which is not covered
by this change.