Fully clone Git repo (instead of a shallow clone)
Unfortunately, when trying to get the commit history for a specific
path, I discovered that the latest commit is returned for each path,
because due to the shallow clone, we only have the latest commit.
Thus, if we want to retrieve the commit history for a path (to get the
date of the last change), we need clone the repository with history.