Update repo information in ES directly after scraping
Previously, the repo information for all scraped repositories was
updated after all repos were scraped. If the scraping of a single
repository failed, none of the repos were updated in ES.

As the repo information itself are only used for the `list-repos`
command so far, this is not a blocker, but still, the behaviour is not
very nice.