Avoid splitting invalid GH repository names
This invalid repo name actually comes from a bug in the tenant config
parser, which results in a repo 'include' that cannot be split into
(owner, /, repo). Thus, the error should be fixed directly in the tenant
parser. Still, I find it's a good idea to make this part more
robust and avoid similar errors in the future.