Don't initialize connections in base command function
This was done to not copy the code in two places. However, initializing
the connections in the base command function will also be done if
somebody used the --help option or provided invalid paramaters for the
scrape command. In those cases the connections would be initialized
before the actual command and its options are evaluated.