Change repo_name field to include Keyword type
This should fix the delete query on outdated repositories. So far this
wasn't working as expected as the query nevery matched a repository due
to a mismatch in the datatype assigned to the repo_name field.

The assigned Text() type by default only make whole words searchable (in
other words it splits by anything that is not a word like slash,
whitespace, comma ...). As the repo_name usually contains a slash when
working with Github, the used term query never matched a repository and
thus no repositories (at least no Github ones) were deleted.