Use Ubuntu 20.04 for Travis CI build
By default Travis CI is using 16.04. This is currently failing because
cryptography cannot be installed without a rust compiler. Most probably
because newer cryptography versions don't provide a wheel for 16.04
anymore.

This also removes the `sudo` key in the Travis CI config as this is
deprecated and not used anymore.