Make GraphQL blame query work without master branch
Currently, the master branch is hardcoded in the GraphQL query that is
used to retrieve the blame information for a file in a GitHub
repository.

Therefore, this query fails in case the repository doesn't provide a
master branch or uses a different default branch name.

This can be fixed by using the defaultBranchRef object which is provided
by the GitHub GraphQL.

To make the whole constrcut a little less error-prone, wrap the
evaluation of the GraphQL response into a try/catch block, so we don't
fail in case the response is not as expected.