Escape special characters in BlockSearch
Currently, there is a bug in Zubbi that causes the search to fail
whenever the query string contains a '/' (or any other "reserved
character" - see [1] for more information).

To fix this, we directly escape those reserved characters internally. We
already do something similar by wrapping the query string into '*' to
enforce a wildcard search every time.

Another idea was to use a simple_query_string [2] instead which provides
a smaller feature set but also doesn't use that many special characters.
However, I didn't manage to get wildcard queries run with that properly,
which would harm the user experience a lot and made searching a lot
harder. So we decided to go for the first solution.

[1]: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html#_reserved_characters
[2]: https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-simple-query-string-query.html#query-dsl-simple-query-string-query