Filter job files for .yaml extension before parsing
Zubbi expects the zuul.d directory to only contain valid yaml files.
Having something like a README.md file in between the job files will
make the parser fail as it doesn't contain valid yaml.

However, having such a file is totally fine for Zuul, so we should
adapt.

As a first step, we filter for `.yaml´ files in the scraper, so that the
parser doesn't need to take care of that.