Fix deprecated Jinja2 imports and update dependencies
This updates some deprecated Jinja2 imports which are shadowed by an
AttributeError raised by the elasticsearch-dsl library. We also add a
test for this specific codepath and add `markupsafe` as explicit
dependency. As a result of the dependency change we also update the
dependencies via pipenv.

As a drive-by fix we also update the `test-requirements.txt` file which
was missing in the latest dependency update as the dependabot only
updates the `pipenv.lock` file.

*Background:*

An error slipped in which results in the description of jobs and roles
not being rendered properly. We didn't spot this problem as the rendered
description is only used within the views of zubbi-web and we don't have
a test for that.

It stated out that the elasticsearch-dsl library is not able to look up
the `description_rendered` attribute of the Block class resulting in the
following exception [1]. However, as this code is only executed within
Jinja2 templates, the exception get swallowed and nothing is displayed.

With this in mindy, the error can be easily reproduced with a few lines
of Python code:

    from zubbi.models import ZuulJob

    job = ZuulJob()
    job.job_name = "foo"
    job.description_html = "<p>Some nice html</p>"
    job.description_rendered

The last line will result in an AttributeError. The important part is
that the exception will only be raised if either `job.description_html`
or `job.description` is set in advance. If none of those is set,
`job.description_rendered` will return None which is OK.

After a longer investigation I found out that the issue is caused by the
latest dependency update [2] which updated also Jinja2. Jinja2 dropped
some deprecated methods [3] in the meantime which are used within the
`description_rendered` property. Thus, fixing those imports also fixed
the AttributeError raised by elasticsearch-dsl.

[1]:

    ---------------------------------------------------------------------------
    KeyError                                  Traceback (most recent call last)
    File ~/.local/share/virtualenvs/zubbi-QuWGSy69/lib/python3.10/site-packages/elasticsearch_dsl/utils.py:159, in AttrDict.__getattr__(self, attr_name)
        158 try:
    --> 159     return self.__getitem__(attr_name)
        160 except KeyError:

    File ~/.local/share/virtualenvs/zubbi-QuWGSy69/lib/python3.10/site-packages/elasticsearch_dsl/utils.py:178, in AttrDict.__getitem__(self, key)
        177 def __getitem__(self, key):
    --> 178     return _wrap(self._d_[key])

    KeyError: 'description_rendered'

    During handling of the above exception, another exception occurred:

    AttributeError                            Traceback (most recent call last)
    Cell In [5], line 1
    ----> 1 job.description_rendered

    File ~/.local/share/virtualenvs/zubbi-QuWGSy69/lib/python3.10/site-packages/elasticsearch_dsl/utils.py:489, in ObjectBase.__getattr__(self, name)
        487 def __getattr__(self, name):
        488     try:
    --> 489         return super(ObjectBase, self).__getattr__(name)
        490     except AttributeError:
        491         f = self.__get_field(name)

    File ~/.local/share/virtualenvs/zubbi-QuWGSy69/lib/python3.10/site-packages/elasticsearch_dsl/utils.py:161, in AttrDict.__getattr__(self, attr_name)
        159     return self.__getitem__(attr_name)
        160 except KeyError:
    --> 161     raise AttributeError(
        162         "{!r} object has no attribute {!r}".format(
        163             self.__class__.__name__, attr_name
        164         )
        165     )

    AttributeError: 'ZuulJob' object has no attribute 'description_rendered'

[2]: https://github.com/bmwcarit/zubbi/pull/83
[3]: https://github.com/pallets/jinja/pull/1544