Fix creation of repo map in TenantParser
The latest change introduced a bug that failed to parse the following
configuration snippets [1] in a tenant-config.yaml file.

This kind of configuration wasn't supported by Zubbi before, but also
didn't fail when parsing the tenant configuration file. For now,
explicitly ignore this kind of configuration.

Additionally, we also add an explicit test for the tenant parser with
a variety of project configs to catch those errors more easily.

This change also fixes the exclusion of jobs from a project, which
didn't work properly due to the usage of "jobs" instead of "job".

[1]:
    # Include nothing from projects foo, bar
    - include: []
      projects:
        - foo
        - bar

    # Exclude jobs and semaphores from foo
    - exclude:
        - job
        - semaphore
      projects:
        - foo