Improve custom Zubbi directives
The new Sphinx version broke our Zubbi directive implementation. The
reason is that the Sphinx log output changed and thus our custom
directives cannot parse the output properly anymore.

We use the Sphinx log to write (and later extract) the custom data from
our directives. This never felt like an appropriate solution but most
propably it was the easiest to do and we didn't find a better solution.

This change improves the Zubbi directives by storing the data in Sphinx'
domaindata dictionary which seems to be more appropriate for this use
case. This is also what other directives (including the "native" Sphinx
directives) do for storing and extracting data).