Make zubbi-web work with path-like role names
Zubbi-scraper is now able to parse roles that are defined in a nested
directory structure. Those roles have a path-like name, e.g. foo/bar.

In order to show those roles properly in zubbi-web, we have to urlencode
the role names in the URL for the details view and also escape the
special characters ("/") in the Elasticsearch query string. The latter
one was already done in other places, so this just aligns the details
query with the "normal" search query.