Switch from pipenv to uv
This sets up the basic uv project structure using a pyproject.toml file
and deletes all pipenv and setuptools related files which are no longer
needed to build the project.

It also updates all dependencies to newer version except elasticsearch
and elasticsearch-dsl which are both pinned to V7. We will update those
in a separate commit.

Using hatchling as build-system allows us to use dynamic versioning
based on git tags, which is not yet supported by uv itself [1].
[1].

Since hatchling includes all files that are not ignored via a .gitignore
file in the source distribution, we are a little more restrictive to
avoid including local and/or temporary files when building the sdist
locally. In case we ever add new files that we want to include which do
not match the pattern, we have to adapt the list.

Compared to the old sdist, the new one is missing the following files
(already exluding the pipenv and setuptools related files which are not
necessary anymore):
```
.coveragerc
.github
.pre-commit-config.yaml
CODEOWNERS
docker
k8s
setup.cfg
zubbi-scraper
```

The generated .whl contains the same files as the one generated via
setuptools (all python files within the zubbi package + static and
template files).

[1]: https://github.com/astral-sh/uv/issues/8714