Run static code checks and tests with uv
This removes tox and runs static code checks and pytest directly with
uv. By using a Makefile we can group commands for different purposes,
so we don't have to run each command manually via uv.

The recommended way with uv is to use the tools subcommand which will
run each tool in a isolated venv. However, this will always run the
latest version of the tool (e.g. pytest), which is not what we want
since we have our dependencies pinned.

The github actions workflow is updated accordingly. Thenew file
structure is based on the examples from the official setup-uv
workflow [1].

[1]: https://github.com/astral-sh/setup-uv